feat(date-logic): default month is now short

+++ packages/date-logic/index.js
@@ -3,7 +3,7 @@
  */
 module.exports = function getDate(
   locale = "en-US",
-  options = { weekday: "long", year: "numeric", month: "long", day: "numeric" }
+  options = { weekday: "long", year: "numeric", month: "short", day: "numeric" }
 ) {
   return new Date().toLocaleDateString(locale, options);
 };
